\name{cqrfactor-package}
\alias{cqrfactor-package}
%\alias{cqrfactor}
\docType{package}
\title{
\packageTitle{cqrfactor}
}
\description{
\packageDescription{cqrfactor}
}
\details{

%The DESCRIPTION file:
%\packageDESCRIPTION{cqrfactor}
%\packageIndices{cqrfactor}
This package computes composite quantile factor and factor loadings for data with
panel structure. The package also offers a function to estimate the number of
factors based on the two information criteria discussed in Huang (2023).
}
\author{
\packageAuthor{cqrfactor}

Maintainer: \packageMaintainer{cqrfactor}
}
\references{
Huang, X. (2023) Composite quantile factor models, working paper.
}

\keyword{  }

