% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mmfactor}
\alias{mmfactor}
\title{Function to compute the composite quantile factors and loadings.}
\usage{
mmfactor(
  y,
  tau,
  q = 5L,
  r = 3L,
  tol = 1e-06,
  maxit = 500L,
  maxit_factor = 500L,
  maxit_loading = 500L,
  convergence = 0L,
  seed = 1L
)
}
\arguments{
\item{y}{A t by n matrix with n variables and t observations each.}

\item{tau}{A vector of quantile positions, and \code{tau = (1:q)/(q+1)},
where q is the numbrer of quantiles.}

\item{q}{The number of quantiles.}

\item{r}{The number of factors.}

\item{tol}{Tolerance for convergence of the MM algorithm. Defaults to 1e-6
when the minconver option is 1. When the minconver option is 0, the
algorithm's convergence depends on the average parameter value change between
iteration steps, and it can take long time to converge. We recommend to use
a bigger number such as 1e-3 when minconver = 1.}

\item{maxit}{The maximum iteration number between factor estimation and factor
loading estimation. Defaults to 500.}

\item{maxit_factor}{The maximum iteration number in the MM algorithm in factor
estimation. Defaults to 500.}

\item{maxit_loading}{The maximum iteration number in the MM algorithm in factor
loading estimation. Defaults to 500.}

\item{convergence}{If the value is 0, the algorithm converges when the minimum
change among all estimated parameters between two iteration steps is less
than or equal to tol. If the value is 1, the algorithm converges when the
average change of all estimated parameters is less than or equal to tol.
This can take a long computation time. If minconver = 1, set tol to a bigger
number such as 1e-3. Defaults to 0.}

\item{seed}{A non-negative integer for random seed used in the C++ function \code{srand} to
initialize factor values. If seed is -1, the initial value of the
factors will be set to the eigenvectors of the outer product of the data
matrix y. Defaults to 1.}
}
\value{
\code{mmfactor} returns a list with the following components:
\item{fmat}{The estimated factors, a t by r matrix, where r is the number
of factors.} \item{lmat}{The estimated factor loading matrix, an n by r
matrix.} \item{quantiles}{The q estimated quantiles of the error term.}
\item{iteration_mat}{A \code{maxit} by 3 matrix. Each row corresponds to an
iteration between factor estimation and factor loading estimation. The
second element is the number of iterations the MM uses for estimating the
factors while the third element is the number of iteration for the factor
loadings.}
}
\description{
This function uses the MM algorithm to compute the composite
quantile factors and factor loadings for a given number of factors.
}
\examples{
\dontrun{

}

}
